# Improve OpenSSL digest performance