--- a/src/native/libs/System.Security.Cryptography.Native/pal_evp.c
+++ b/src/native/libs/System.Security.Cryptography.Native/pal_evp.c
@@ -12,33 +12,51 @@
 
 #ifdef NEED_OPENSSL_3_0
 #define SETUP_MD_FETCH(export, fn, name, query) \
-    static const EVP_MD* g_evpFetch##name = NULL; \
-    static pthread_once_t g_evpFetchInit##name = PTHREAD_ONCE_INIT; \
-    static void EnsureFetchEvpMd##name(void) \
+    static const EVP_MD* g_evpFetch##export = NULL; \
+    static pthread_once_t g_evpFetchInit##export = PTHREAD_ONCE_INIT; \
+    static void EnsureFetchEvpMd##export(void) \
     { \
         if (API_EXISTS(EVP_MD_fetch)) \
         { \
             ERR_clear_error(); \
-            g_evpFetch##name = EVP_MD_fetch(NULL, #name, query); \
+            g_evpFetch##export = EVP_MD_fetch(NULL, name, query); \
         } \
 \
-        if (g_evpFetch##name == NULL) \
+        if (g_evpFetch##export == NULL) \
         { \
-            g_evpFetch##name = fn(); \
+            g_evpFetch##export = fn(); \
         } \
     } \
     \
-    const EVP_MD* export(void) \
+    const EVP_MD* CryptoNative_##export(void) \
     { \
-        pthread_once(&g_evpFetchInit##name, EnsureFetchEvpMd##name); \
-        return g_evpFetch##name; \
+        pthread_once(&g_evpFetchInit##export, EnsureFetchEvpMd##export); \
+        return g_evpFetch##export; \
     }
+#define SETUP_MD_FETCH_LIGHTUP_SHA3(...) SETUP_MD_FETCH(__VA_ARGS__)
 #else
-#define SETUP_MD_FETCH(export, fn, name) \
-    const EVP_MD* export(void)
+#define SETUP_MD_FETCH(export, fn, name, query) \
+    const EVP_MD* CryptoNative_##export(void) \
     { \
         return fn(); \
     }
+#if HAVE_OPENSSL_SHA3
+#define SETUP_MD_FETCH_LIGHTUP_SHA3(export, fn, name, query) \
+    const EVP_MD* CryptoNative_##export(void) \
+    { \
+        if (API_EXISTS(fn)) \
+        { \
+            return fn(); \
+        } \
+        \
+        return NULL; \
+    }
+#else
+    const EVP_MD* CryptoNative_##export(void) \
+    { \
+        return NULL; \
+    }
+#endif
 #endif
 
 
@@ -299,76 +317,16 @@ int32_t CryptoNative_EvpMdSize(const EVP_MD* md)
 
 // MD5 should use a non-FIPS implementation if it is available. We should not fail
 // to fetch MD5 even on a FIPS enforced system.
-SETUP_MD_FETCH(CryptoNative_EvpMd5, EVP_md5, MD5, "-fips")
-SETUP_MD_FETCH(CryptoNative_EvpSha1, EVP_sha1, SHA1, NULL)
-SETUP_MD_FETCH(CryptoNative_EvpSha256, EVP_sha256, SHA256, NULL)
-SETUP_MD_FETCH(CryptoNative_EvpSha384, EVP_sha384, SHA384, NULL)
-SETUP_MD_FETCH(CryptoNative_EvpSha512, EVP_sha512, SHA512, NULL)
-
-const EVP_MD* CryptoNative_EvpSha3_256(void)
-{
-    // No error queue impact.
-#if HAVE_OPENSSL_SHA3
-    if (API_EXISTS(EVP_sha3_256))
-    {
-        return EVP_sha3_256();
-    }
-#endif
-
-    return NULL;
-}
-
-const EVP_MD* CryptoNative_EvpSha3_384(void)
-{
-    // No error queue impact.
-#if HAVE_OPENSSL_SHA3
-    if (API_EXISTS(EVP_sha3_384))
-    {
-        return EVP_sha3_384();
-    }
-#endif
-
-    return NULL;
-}
-
-const EVP_MD* CryptoNative_EvpSha3_512(void)
-{
-    // No error queue impact.
-#if HAVE_OPENSSL_SHA3
-    if (API_EXISTS(EVP_sha3_512))
-    {
-        return EVP_sha3_512();
-    }
-#endif
-
-    return NULL;
-}
-
-const EVP_MD* CryptoNative_EvpShake128(void)
-{
-    // No error queue impact.
-#if HAVE_OPENSSL_SHA3
-    if (API_EXISTS(EVP_shake128))
-    {
-        return EVP_shake128();
-    }
-#endif
-
-    return NULL;
-}
-
-const EVP_MD* CryptoNative_EvpShake256(void)
-{
-    // No error queue impact.
-#if HAVE_OPENSSL_SHA3
-    if (API_EXISTS(EVP_shake256))
-    {
-        return EVP_shake256();
-    }
-#endif
-
-    return NULL;
-}
+SETUP_MD_FETCH(EvpMd5, EVP_md5, "MD5", "-fips")
+SETUP_MD_FETCH(EvpSha1, EVP_sha1, "SHA1", NULL)
+SETUP_MD_FETCH(EvpSha256, EVP_sha256, "SHA256", NULL)
+SETUP_MD_FETCH(EvpSha384, EVP_sha384, "SHA384", NULL)
+SETUP_MD_FETCH(EvpSha512, EVP_sha512, "SHA512", NULL)
+SETUP_MD_FETCH_LIGHTUP_SHA3(EvpSha3_256, EVP_sha3_256, "SHA3-256", NULL)
+SETUP_MD_FETCH_LIGHTUP_SHA3(EvpSha3_384, EVP_sha3_384, "SHA3-384", NULL)
+SETUP_MD_FETCH_LIGHTUP_SHA3(EvpSha3_512, EVP_sha3_512, "SHA3-512", NULL)
+SETUP_MD_FETCH_LIGHTUP_SHA3(EvpShake128, EVP_shake128, "SHAKE-128", NULL)
+SETUP_MD_FETCH_LIGHTUP_SHA3(EvpShake256, EVP_shake256, "SHAKE-256", NULL)
 
 int32_t CryptoNative_GetMaxMdSize(void)
 {