--- a/src/native/libs/System.Security.Cryptography.Native/pal_evp.c
+++ b/src/native/libs/System.Security.Cryptography.Native/pal_evp.c
@@ -28,7 +28,7 @@
         } \
     } \
     \
-    const EVP_MD* CryptoNative_##export(void) \
+    const EVP_MD* export(void) \
     { \
         pthread_once(&g_evpFetchInit##export, EnsureFetchEvpMd##export); \
         return g_evpFetch##export; \
@@ -36,13 +36,13 @@
 #define SETUP_MD_FETCH_LIGHTUP_SHA3(...) SETUP_MD_FETCH(__VA_ARGS__)
 #else
 #define SETUP_MD_FETCH(export, fn, name, query) \
-    const EVP_MD* CryptoNative_##export(void) \
+    const EVP_MD* export(void) \
     { \
         return fn(); \
     }
 #if HAVE_OPENSSL_SHA3
 #define SETUP_MD_FETCH_LIGHTUP_SHA3(export, fn, name, query) \
-    const EVP_MD* CryptoNative_##export(void) \
+    const EVP_MD* export(void) \
     { \
         if (API_EXISTS(fn)) \
         { \
@@ -52,7 +52,7 @@
         return NULL; \
     }
 #else
-    const EVP_MD* CryptoNative_##export(void) \
+    const EVP_MD* export(void) \
     { \
         return NULL; \
     }
@@ -317,16 +317,16 @@ int32_t CryptoNative_EvpMdSize(const EVP_MD* md)
 
 // MD5 should use a non-FIPS implementation if it is available. We should not fail
 // to fetch MD5 even on a FIPS enforced system.
-SETUP_MD_FETCH(EvpMd5, EVP_md5, "MD5", "-fips")
-SETUP_MD_FETCH(EvpSha1, EVP_sha1, "SHA1", NULL)
-SETUP_MD_FETCH(EvpSha256, EVP_sha256, "SHA256", NULL)
-SETUP_MD_FETCH(EvpSha384, EVP_sha384, "SHA384", NULL)
-SETUP_MD_FETCH(EvpSha512, EVP_sha512, "SHA512", NULL)
-SETUP_MD_FETCH_LIGHTUP_SHA3(EvpSha3_256, EVP_sha3_256, "SHA3-256", NULL)
-SETUP_MD_FETCH_LIGHTUP_SHA3(EvpSha3_384, EVP_sha3_384, "SHA3-384", NULL)
-SETUP_MD_FETCH_LIGHTUP_SHA3(EvpSha3_512, EVP_sha3_512, "SHA3-512", NULL)
-SETUP_MD_FETCH_LIGHTUP_SHA3(EvpShake128, EVP_shake128, "SHAKE-128", NULL)
-SETUP_MD_FETCH_LIGHTUP_SHA3(EvpShake256, EVP_shake256, "SHAKE-256", NULL)
+SETUP_MD_FETCH(CryptoNative_EvpMd5, EVP_md5, "MD5", "-fips")
+SETUP_MD_FETCH(CryptoNative_EvpSha1, EVP_sha1, "SHA1", NULL)
+SETUP_MD_FETCH(CryptoNative_EvpSha256, EVP_sha256, "SHA256", NULL)
+SETUP_MD_FETCH(CryptoNative_EvpSha384, EVP_sha384, "SHA384", NULL)
+SETUP_MD_FETCH(CryptoNative_EvpSha512, EVP_sha512, "SHA512", NULL)
+SETUP_MD_FETCH_LIGHTUP_SHA3(CryptoNative_EvpSha3_256, EVP_sha3_256, "SHA3-256", NULL)
+SETUP_MD_FETCH_LIGHTUP_SHA3(CryptoNative_EvpSha3_384, EVP_sha3_384, "SHA3-384", NULL)
+SETUP_MD_FETCH_LIGHTUP_SHA3(CryptoNative_EvpSha3_512, EVP_sha3_512, "SHA3-512", NULL)
+SETUP_MD_FETCH_LIGHTUP_SHA3(CryptoNative_EvpShake128, EVP_shake128, "SHAKE-128", NULL)
+SETUP_MD_FETCH_LIGHTUP_SHA3(CryptoNative_EvpShake256, EVP_shake256, "SHAKE-256", NULL)
 
 int32_t CryptoNative_GetMaxMdSize(void)
 {