@bartonjs 2025-08-11
Wouldn't we get basically the same performance improvement by just caching the answer in managed and not doing anything here in native?