@vcsjones 2025-08-11
> Wouldn't we get basically the same performance improvement by just caching the answer in managed and not doing anything here in native?

I don't think so.

I attempted to address that here:

> Even though we memoize the value over on the managed side in the HashAlgorithmDispenser, this memorization only really avoids the p/invoke boundary. What is actually getting memoized is a lookup function (more or less)

We are already doing the memoization. For example:

https://github.com/dotnet/runtime/blob/f30c40e179a66aa6f386b649c05a79ddd365bb61/src/libraries/Common/src/Interop/Unix/System.Security.Cryptography.Native/Interop.EVP.DigestAlgs.cs#L43-L44

The problem is what `EVP_sha256` returns on OpenSSL is another lazy lookup itself for EVP_fetch.