@bartonjs 2025-08-11
Weird, it looks like stable info to me:

https://github.com/openssl/openssl/blob/076f7b24fee1b80a5cda898f385ae813217c823f/crypto/evp/legacy_sha.c#L92-L105

But perhaps EVP_ORIG_GLOBAL means something like "ask fetch first, but use the LEGACY_EVP_MD_METH_TABLE as a fallback"