@vcsjones 2025-08-13
/ba-g failures are unrelated, and tracked elsewhere.